Crisis averted

+++ controller/graph.js
@@ -16,6 +16,25 @@ const generateQRCode = async (link) => {
 
 const uploadGraph = async (req, res) => {
     console.log(req.body);
+
+    const nodes = req.body.nodes;
+    const edges = req.body.edges;
+
+    // Get all node ids that are not connected to any edge
+    const nodeIds = nodes.map((node) => node.id);
+    const edgeNodeIds = edges.flatMap((edge) => [edge.nodeA, edge.nodeB]);
+    const unconnectedNodeIds = nodeIds.filter((id) => !edgeNodeIds.includes(id));
+
+    // Return with an error if there are unconnected nodes
+    if (unconnectedNodeIds.length > 0) {
+        return res.status(400).json({
+            error: "Graph contains unconnected nodes",
+            unconnectedNodeIds: unconnectedNodeIds,
+        });
+    }
+
+
+
     // Drop every node and edge
     try {
         await prisma.node.deleteMany();
